requirements.txt
anthropic>=0.25.0
python-telegram-bot>=20.7
tweepy>=4.14.0
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
beautifulsoup4>=4.12.0
